recommendation_engine/requirements.txt
flask==2.0.1
numpy==1.21.0
pandas==1.3.0
scikit-learn==0.24.2
mysql-connector-python==8.0.26
python-dotenv==0.19.0
gunicorn==20.1.0
